Make transliteration functions customizable
Transliteration can be used by the Aligner and DocSplitter for getting better
alignments and split positions. Ultimately the transliteration functions
should not have to come from this the textalign module, but can/should come
from elsewhere.
Moving towards that, I have removed the hard-coded or default function values where
a transliteration function was used and replaced it with an Optional[Callable] attribute,
that can be any transliteration function.
Now the explicit transliteration function (currently "unidecode_ger") is only referenced in
the texts.
There are still some TODOs:
* Check whether there are specific constraints on the translit function,
specifically whether the transliterated text has to be of the same length
as the original text
* Currently the unidecode_ger function is flawed, resulting in one of the tests
in test_translit.py to fail