Update translit escape functions
* Trema ("\u0308") is now escaped/unescaped so that it is retained
during unidecode_ger (previously it was removed by unidecode).
* This helps keeping the strings the same lengths during transliteration
instead of changing the length. Altered lengths would be a problem during
document splitting because the input files contain character offsets, that
would be corrupted if the string lengths change during processing.
* Note that, because the "combining small letter e" (\u0364) is converted to
the trema (\u0308), transliterated text contains two representations for umlauts,
namely (1) the unified symbols ("ä","ü","ö", etc.) for Umlauts that where present in
the original text as a single character, and (2) the ASCII letter plus trema
(e.g. "a\u0308") for Umlauts expressed as a combination of the letter and the combining
small e in the original text.